Django nf-Admin support, and other minor bug cleanup (thanks for the patch!)
*WARNING*: the schema has changed with this checkin.

Added an advanced option 'path_type' which is for restricting the 
selection marking on the navtree and navbars. 

Rational:

If you have two entries on the navtree which are sub-urls of the request, 
and they are under different sub-trees, then both will be selected. 
If you have a 'home' entry for '/', then it will always be selected. 
In the case of navbars, this may mean that only the top level navbar is
rendered (which was the case before this change and a database update).

It is also not always the wanted behavior to have an entry selected if the 
url is actually  below that point. As such the following advanced 
'path selection types' are now available:

'None' - This is an early out for a navigation entry which you never want 
         to show up selected. I actually have a complex use case for this.
'Exact' - Match the url exactly. So 'home' will only be selected if the url
          is '/' exactly.
'ExactOrParent' - Match the url exactly, or if a sub tree entry is selected.
'OnPathOrParent' - This is the default subpath matching and parents behavior.
 


git-svn-id: http://django-navbar.googlecode.com/svn/trunk@8 40ae0140-9c48-0410-9b1c-13f58143d509